Several fixes
Updated gg_version check to give the version numbers in descending order, and to include all valid version numbers, even if they are not found in the symbol_path.

Also added the ability for the symbols to be copied from \\qnapnas.graphon.com\builds (or whatever the user sets the backup path to, though it may not work as well unless the structure of the directory is the same).